minor adjustments. Added new functionality Rate Calculator depends on Type rate & payment Manager

+++ src/main/java/carsharing/web/controller/CardController.java
@@ -1,16 +1,20 @@
 package carsharing.web.controller;
 
+import carsharing.dao.model.Customer;
 import carsharing.service.CardService;
 import carsharing.web.dto.CardDTO;
 import carsharing.web.mapper.CardMapper;
 import org.mapstruct.factory.Mappers;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
 @RestController
-@RequestMapping("api/card")
+@RequestMapping("api/cards")
 public class CardController {
 
     @Autowired
@@ -19,19 +23,22 @@ public class CardController {
     private CardMapper cardMapper = Mappers.getMapper(CardMapper.class);
 
     @GetMapping(value = "/list")
-    public List<CardDTO> getOwnCardList() {
-        return cardMapper.convertToDTO(cardService.getAll());
+    public List<CardDTO> getCardList(@AuthenticationPrincipal Customer customerDetails) {
+        return cardMapper.convertToDTO(customerDetails.getCards());
     }
 
     @GetMapping(value = "/{id}/delete")
-    public String deleteCard(@PathVariable("id") Long id) {
-        cardService.delete(id);
-        return "Card delete - " + id;
+    public void deleteCard(@PathVariable("id") Long id, @AuthenticationPrincipal Customer customerDetails) {
+        if (customerDetails.getCards().contains(cardService.findById(id))) {
+            cardService.delete(id);
+        }
+        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Denied, not enough rights");
     }
 
     @PostMapping(value = "/add", produces = "application/json", consumes="application/json")
-    public String addCard(@RequestBody CardDTO cardDTO) {
-        cardService.create(cardMapper.convertToEntity(cardDTO));
-        return "Card added";
+    public CardDTO addCard(@RequestBody CardDTO cardDTO, @AuthenticationPrincipal Customer customerDetails) {
+        cardDTO.setCustomer(customerDetails);
+        cardService.save(cardMapper.convertToEntity(cardDTO));
+        return cardDTO;
     }
 }
